Cancel entity damage if not authenticated by authme
We should cancel the damage no matter what if either of the players are not authenticated by AuthMe. An un-logged-in player cannot do damage or receive it, and should never tag or get tagged.

I should have moved the attacker variable up, and should have checked if AuthMe is available. But I didn't want to interfere with your coding style. Just trying to help a bit.